requirements.txt
streamlit
torch
transformers
nltk
sacremoses
pandas
regex
mock
mosestokenizer
bitsandbytes
scipy
accelerate
datasets
sentencepiece
speechbrain
fastapi
uvicorn
httpx
python-multipart
python-dotenv
groq
langchain
langchain-community
langchain-groq
gTTS==2.5.4